services/agents/voice/pyproject.toml
[tool.poetry]
name = "voice-agent"
version = "1.0.0"
description = "A.R.C. Framework's designated front-of-the-line service for all real-time, bidirectional voice and text interactions."
authors = ["A.R.C. Engineering <engineering@arc.framework>"]
readme = "README.md"
packages = [{include = "voice_agent", from = "src"}]

[tool.poetry.dependencies]
# We are explicitly locking to Python 3.12 to ensure stability and avoid compilation errors.
python = ">=3.12, <3.13"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = "^2.7.1"
pipecat-ai = "^0.0.94"
pulsar-client = {extras = ["all"], version = "^3.5.0"}

# The SDK is lightweight and does not cause the conflict.
# The exporter libraries have been removed to solve the dependency issue.
opentelemetry-sdk = "^1.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
